Optimize display-buffer-alist for Go
Tested these scenarios:

Starting with a single window, press C-x g (magit-status). This splits the window and shows magit on the right. Click on
a Go file in the magit window, now the Go file should be open in the original window, not the magit window. Press q to
close magit and get back to single window.

Starting with a single window, it is possible to split it in two and get the same Go buffer in both windows. This is
useful for editing multiple sections of the same file.

When multiple Go windows are open, C-x f opens new Go files in the current window.

When in vterm, C-x f opens Go files in one of the existing windows dedicated to Go files.