Make git-link work over go mod cache
Frequently definition lookup in Go takes you to a Go file in go mod cache; this is technically not a Git repository but
sometimes packages have github.com in them. In this case the replacement for git-link will now compute the GitHub link
as if this was a local copy of the repository.

Regions and fancy arguments are not yet supported.